feature: (WIP) run commands on server on disconnect (for presence)
This change adds the ability to defer some commands to be run when the client
is disconnected. This can be used to implement presence (insert a row into a
table after connecting, tell the server to remove that row when your connection
goes away.)

This was implemented on the server side by adding a new flag to write ops:
on_disconnect. If set to true, the op is deferred until the client disconnects.

On the client, it was implemented by adding a .onDisconnect() method to Collection.
That returns a new Collection with a flag set that will set the on_disconnect flag
in requests.

Still to do:

 * Tests
 * Make sure on_disconnect requests are properly cleaned up after running

Future improvements:

 * Reliability. The current implementation will lose onDisconnect events if
   the process handling the client request dies for whatever reason.
   (As a side note, I've had orphaned onDisconnect calls happen in Firebase too,
   so I suspect they have a similar implementation).

   You could solve this by writing serialized onDisconnect requests into
   another rdb table, along with identifying information about the process
   handling it. Processes would then ping regularly into another table to show
   they are still running. An auditor process would then run regularly and
   detect down processes, and run their onDisconnect requests. There are all
   sorts of complexities and race conditions around a system like this.

 * One nice thing you can do with the onDisconnect feature is leave behind a
   "last online at" timestamp on a user record or something. In order to do this,
   we'd need something similar to Firebase's SERVER_TIMESTAMP placeholder for
   the delayed write request.

 * Canceling the pending action. This is already supported server-side, but
   needs support in the frontend client.